refactor(og): split og-palette.ts for client bundle isolation (plan022)
og.ts 가 node:fs import 하므로 client component (Avatar.tsx) 에서 직접 import 시
Turbopack 이 fs 를 클라이언트 번들에 포함하려다 실패. palette 데이터를
pure module (og-palette.ts) 로 분리, og.ts 는 re-export 만 유지.

scope: plan022 의 Avatar 가 OG_CATEGORY_HEX 단일 소스를 재사용해야 하는데,
client component 라 분리 없이는 빌드 깨짐. 의도된 scope 확장 (executor 보고 → ACCEPT).
