+++ src/lib/og.ts
@@ -71,42 +71,7 @@ export const OG_LAYOUT = {
   brandBarHeight: 4,
 } as const;
 
-/**
- * plan009 카테고리 토큰의 satori 호환 hex 매핑.
- * 원본 oklch(0.74 0.09 H) — sRGB 변환은 culori(formatHex)로 사전 계산.
- * satori 는 oklch 미지원이라 직접 hex 박아 넣음.
- *
- * 키는 정규화된 짧은 alias 기준. categoryIcons 의 풀네임(database/javascript)은
- * OG_CATEGORY_ALIAS 로 흡수. 미등록 카테고리(architecture/finance/git/go/html/http/internet/
- * interview/kafka/network/redis/resume/css/기술공유 등)는 OG_CATEGORY_DEFAULT_HEX(brand teal)
- * 로 fallback — 향후 plan 에서 도메인 친화 색을 점진 확장 예정.
- */
-export const OG_CATEGORY_HEX: Record<string, string> = {
-  ai: "#a4a3e2",        // oklch(0.74 0.09 285)
-  algorithm: "#de958e", // oklch(0.74 0.09 25)
-  db: "#d79c73",        // oklch(0.74 0.09 55)
-  devops: "#87ba88",    // oklch(0.74 0.09 145)
-  java: "#64bead",      // oklch(0.74 0.09 180)
-  js: "#c1a966",        // oklch(0.74 0.09 90)
-  react: "#64b8d2",     // oklch(0.74 0.09 220)
-};
-
-export const OG_CATEGORY_DEFAULT_HEX = "#3fbac9"; // brand-400 fallback
-
-/**
- * categoryIcons(src/infra/db/constants.ts) 의 풀네임을 짧은 alias 로 매핑.
- * lowercase 변환 후 1차 lookup → 매치 시 alias 키로 OG_CATEGORY_HEX 재조회.
- */
-const OG_CATEGORY_ALIAS: Record<string, string> = {
-  database: "db",
-  javascript: "js",
-};
-
-export function getCategoryHex(category: string): string {
-  const raw = category.toLowerCase();
-  const key = OG_CATEGORY_ALIAS[raw] ?? raw;
-  return OG_CATEGORY_HEX[key] ?? OG_CATEGORY_DEFAULT_HEX;
-}
+export { OG_CATEGORY_HEX, OG_CATEGORY_DEFAULT_HEX, getCategoryHex } from "@/lib/og-palette";
 
 const HEX_PATTERN = /^#[0-9a-fA-F]{6}$/;
